requirements.txt
fastapi
uvicorn
httpx
openai
pydantic
python-dotenv
solders
base58
